# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: divideIt API
  description: |
    RESTful API for splitting videos into random segments. Perfect for creating content for Reels, TikTok, and YouTube Shorts.
    
    ## Features
    - Upload and split videos in a single request
    - Support for MP4, MOV, and AVI formats
    - Random segment generation without AI
    - Customizable segment count and duration
    - Download individual segments
    
    ## Rate Limiting
    - General API: 100 requests per 15 minutes per IP
    - Upload endpoint: 10 requests per hour per IP
    
    ## File Size Limits
    - Maximum file size: 500MB
    - Supported formats: MP4, MOV, AVI
  version: 1.0.0
  contact:
    name: divideIt Support
    url: https://github.com/your-username/divideIt
    email: support@divideit.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.divideit.example.com/api
    description: Production server

tags:
  - name: Health
    description: Health check endpoints
  - name: Videos
    description: Video upload, splitting, and download operations

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Returns the health status of the API server
      operationId: getHealth
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              example:
                status: 'ok'
                timestamp: '2024-01-15T10:30:00.000Z'
                uptime: 3600
        '500':
          $ref: '#/components/responses/ServerError'

  /videos/split:
    post:
      tags:
        - Videos
      summary: Upload and split video
      description: |
        Uploads a video file and splits it into random segments in a single request.
        
        The video will be processed using FFmpeg to extract segments based on the provided parameters.
        Segments are generated randomly within the specified duration constraints.
        
        **Processing Time**: Depends on video size and number of segments. Typically 1-5 minutes for a 100MB video.
      operationId: splitVideo
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - video
              properties:
                video:
                  type: string
                  format: binary
                  description: Video file (MP4, MOV, or AVI format, max 500MB)
                segmentCount:
                  type: integer
                  minimum: 1
                  maximum: 20
                  default: 5
                  description: Number of segments to generate
                minSegmentDuration:
                  type: number
                  minimum: 1
                  maximum: 300
                  default: 5
                  description: Minimum duration of each segment in seconds
                maxSegmentDuration:
                  type: number
                  minimum: 1
                  maximum: 300
                  default: 60
                  description: Maximum duration of each segment in seconds
            encoding:
              video:
                contentType: video/mp4, video/quicktime, video/x-msvideo
      responses:
        '200':
          description: Video successfully split into segments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SplitVideoResponse'
              example:
                success: true
                data:
                  originalVideo:
                    filename: 'my-video.mp4'
                    duration: 120.5
                    metadata:
                      duration: 120.5
                      width: 1920
                      height: 1080
                      format: 'mov,mp4,m4a,3gp,3g2,mj2'
                      size: 52428800
                  segments:
                    - segmentNumber: 1
                      startTime: 10.5
                      endTime: 25.3
                      duration: 14.8
                      downloadUrl: '/api/videos/download/segment_1_a1b2c3d4.mp4'
                    - segmentNumber: 2
                      startTime: 45.2
                      endTime: 78.9
                      duration: 33.7
                      downloadUrl: '/api/videos/download/segment_2_e5f6g7h8.mp4'
                  totalSegments: 2
        '400':
          description: Bad request - Invalid file or parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidFile:
                  value:
                    success: false
                    error:
                      message: 'Invalid file type. Only MP4, MOV, and AVI files are allowed.'
                invalidDuration:
                  value:
                    success: false
                    error:
                      message: 'Video duration (3s) is less than minimum segment duration (5s)'
                validationError:
                  value:
                    success: false
                    error:
                      message: 'Validation error: segmentCount must be between 1 and 20'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many requests - Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'

  /videos/download/{filename}:
    get:
      tags:
        - Videos
      summary: Download video segment
      description: Downloads a processed video segment file
      operationId: downloadSegment
      parameters:
        - name: filename
          in: path
          required: true
          description: Filename of the segment to download (e.g., segment_1_a1b2c3d4.mp4)
          schema:
            type: string
            example: segment_1_a1b2c3d4.mp4
      responses:
        '200':
          description: Segment file
          content:
            video/mp4:
              schema:
                type: string
                format: binary
        '404':
          description: Segment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'

components:
  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: 'ok'
        timestamp:
          type: string
          format: date-time
          example: '2024-01-15T10:30:00.000Z'
        uptime:
          type: number
          description: Server uptime in seconds
          example: 3600

    VideoMetadata:
      type: object
      properties:
        duration:
          type: number
          description: Video duration in seconds
          example: 120.5
        width:
          type: integer
          description: Video width in pixels
          example: 1920
        height:
          type: integer
          description: Video height in pixels
          example: 1080
        format:
          type: string
          description: Video container format
          example: 'mov,mp4,m4a,3gp,3g2,mj2'
        size:
          type: integer
          description: File size in bytes
          example: 52428800

    VideoSegment:
      type: object
      properties:
        segmentNumber:
          type: integer
          description: Sequential segment number (1-indexed)
          example: 1
        startTime:
          type: number
          description: Start time in seconds
          example: 10.5
        endTime:
          type: number
          description: End time in seconds
          example: 25.3
        duration:
          type: number
          description: Segment duration in seconds
          example: 14.8
        downloadUrl:
          type: string
          description: URL to download the segment
          example: '/api/videos/download/segment_1_a1b2c3d4.mp4'

    OriginalVideo:
      type: object
      properties:
        filename:
          type: string
          description: Original filename
          example: 'my-video.mp4'
        duration:
          type: number
          description: Video duration in seconds
          example: 120.5
        metadata:
          $ref: '#/components/schemas/VideoMetadata'

    SplitVideoResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            originalVideo:
              $ref: '#/components/schemas/OriginalVideo'
            segments:
              type: array
              items:
                $ref: '#/components/schemas/VideoSegment'
            totalSegments:
              type: integer
              description: Total number of segments created
              example: 5

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
              description: Error message
              example: 'Invalid file type. Only MP4, MOV, and AVI files are allowed.'
            code:
              type: string
              description: Error code (optional)
              example: 'INVALID_FILE_TYPE'

  responses:
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              message: 'An internal server error occurred'

  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (future feature)